fix: homepage actions
