+++ src/modules/project/sagas.ts
@@ -32,7 +32,7 @@ import { api } from 'lib/api'
 import { Project } from 'modules/project/types'
 import { Scene } from 'modules/scene/types'
 import { getData as getProjects } from 'modules/project/selectors'
-import { getData as getScenes, getScene } from 'modules/scene/selectors'
+import { getScene } from 'modules/scene/selectors'
 import { EMPTY_SCENE_METRICS } from 'modules/scene/constants'
 import { createScene, setGround, applyLayout } from 'modules/scene/actions'
 import { SET_EDITOR_READY, setEditorReady, takeScreenshot, setExportProgress, createEditorScene } from 'modules/editor/actions'
@@ -43,6 +43,7 @@ import { AUTH_SUCCESS, AuthSuccessAction } from 'modules/auth/actions'
 import { createFiles } from './export'
 import { didUpdateLayout } from './utils'
 import { getSub } from 'modules/auth/selectors'
+import { getSceneByProjectId } from 'modules/scene/utils'
 
 const DEFAULT_GROUND_ASSET: Asset = {
   id: 'da1fed3c954172146414a66adfa134f7a5e1cb49c902713481bf2fe94180c2cf',
@@ -113,9 +114,7 @@ function* handleCreateProjectFromTemplate(action: CreateProjectFromTemplateActio
 function* handleDuplicateProject(action: DuplicateProjectAction) {
   const { project } = action.payload
 
-  const scenes: ReturnType<typeof getScenes> = yield select(getScenes)
-  const scene = scenes[project.sceneId]
-  if (!scene) return
+  const scene = yield getSceneByProjectId(project.id)
 
   const newScene = { ...scene, id: uuidv4() }
   const newProject = { ...project, sceneId: newScene.id, id: uuidv4(), createdAt: new Date().toISOString() }
@@ -152,7 +151,7 @@ function* handleEditProject(action: EditProjectAction) {
 
 function* handleExportProject(action: ExportProjectRequestAction) {
   const { project } = action.payload
-  const scene: Scene = yield select(getScene(project.sceneId))
+  const scene = yield getSceneByProjectId(project.id)
 
   let zip = new JSZip()
   let sanitizedName = project.title.replace(/\s/g, '_')
